requirements.txt
# Core dependencies for topic modeling and clustering
bertopic>=0.15.0
openai>=1.0.0
datasets>=2.0.0
datamapplot>=0.1.0

# Scientific computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Dimensionality reduction and clustering
umap-learn>=0.5.0
hdbscan>=0.8.0

# Visualization
matplotlib>=3.5.0
plotly>=5.0.0
seaborn>=0.11.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Progress bars
tqdm>=4.60.0

# Environment variables
python-dotenv>=0.19.0

# Streamlit and related
streamlit>=1.20.0
streamlit-autorefresh>=0.0.1

# Google Sheets integration
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
oauth2client>=4.1.3

# Enhanced error handling
backoff>=2.0.0
retrying>=1.3.3

# File upload widget for Jupyter
ipyfilechooser>=0.6.0
